Code fixes
