+++ cartridges/app_storefront_pdp_extend/cartridge/scripts/helpers/contentAssetHelpers.js
@@ -1,0 +1,15 @@
+'use strict';
+
+function getClientContentAsset(contentAssetID) {
+    var ContentMgr = require('dw/content/ContentMgr');
+    var customerContent = ContentMgr.getContent(contentAssetID);
+    if (customerContent && customerContent.online) {
+        return customerContent;
+    } else {
+        return null;
+    }
+}
+
+module.exports = {
+    getClientContentAsset: getClientContentAsset
+}
