+++ cartridges/app_storefront_pdp_extend/cartridge/scripts/helpers/notificationHelpers.js
@@ -1,26 +1,10 @@
 'use strict';
 
-function getClientContentAsset(contentAssetID) {
-    var ContentMgr = require('dw/content/ContentMgr');
-    var customerContent = ContentMgr.getContent(contentAssetID);
-    return customerContent;
-}
-
-function getCustomObject(customObjectType, primaryKey) {
-    var CustomObjectMgr = require('dw/object/CustomObjectMgr');
-
-    return CustomObjectMgr.getCustomObject(customObjectType, primaryKey);
-}
-
-function createCustomObject(customObjectType, primaryKey) {
-    var CustomObjectMgr = require('dw/object/CustomObjectMgr');
-
-    return CustomObjectMgr.createCustomObject(customObjectType, primaryKey);
-}
+var customObjectHelper = require('*/cartridge/scripts/helpers/customObjectHelpers.js');
 
 function addSubscriptionEntry(type, primaryKey, phoneNumber, productName) {
 
-    var subscriptionEntry = getCustomObject(type, primaryKey);
+    var subscriptionEntry = customObjectHelper.getCustomObject(type, primaryKey);
 
     if (subscriptionEntry) {
         var subsArray = subscriptionEntry.custom.phoneNumbers;
@@ -31,16 +15,14 @@ function addSubscriptionEntry(type, primaryKey, phoneNumber, productName) {
         phoneNumbersArray.push(phoneNumber);
         subscriptionEntry.custom.phoneNumbers = phoneNumbersArray
     } else {
-        subscriptionEntry = createCustomObject(type, primaryKey);
+        subscriptionEntry = customObjectHelper.createCustomObject(type, primaryKey);
         subscriptionEntry.custom.phoneNumbers = new Array(phoneNumber);
         subscriptionEntry.custom.productName = productName
     }
 
     return subscriptionEntry;
-
 }
 
 module.exports = {
     addSubscriptionEntry: addSubscriptionEntry,
-    getClientContentAsset: getClientContentAsset
 }
