+++ cartridges/app_storefront_pdp_extend/cartridge/templates/default/product/components/addToCartProduct.isml
@@ -12,7 +12,7 @@
     </div>
 
     <div class="col-sm-12 text-left ${product.availabilityModel.inStock ? "d-none" : ""}">
-        <isinclude url="${URLUtils.url('NotifySubscription-Show', 'pid', product.id, 'productName',product.productName,'productAvailability',product.available)}" />
+        <isinclude url="${URLUtils.url('NotifySubscription-Show', 'pid', product.id)}" />
     </div>
 
     <isinclude template="product/components/addToCartButtonExtension" />
