+++ site_template/custom-objecttype-definitions.xml
@@ -35,56 +35,4 @@
             </attribute-group>
         </group-definitions>
     </custom-type>
-
-    <custom-type type-id="Product-Subscribe-Notification">
-        <display-name xml:lang="x-default">productID</display-name>
-        <staging-mode>no-staging</staging-mode>
-        <storage-scope>organization</storage-scope>
-        <key-definition attribute-id="productID">
-            <type>string</type>
-            <min-length>0</min-length>
-        </key-definition>
-        <attribute-definitions>
-            <attribute-definition attribute-id="PhoneNumberAndCustomerName">
-                <display-name xml:lang="x-default">Phone Number And Customer Name</display-name>
-                <type>set-of-string</type>
-                <localizable-flag>false</localizable-flag>
-                <mandatory-flag>false</mandatory-flag>
-                <externally-managed-flag>false</externally-managed-flag>
-            </attribute-definition>
-            <attribute-definition attribute-id="clientNumber">
-                <display-name xml:lang="x-default">clientNumber</display-name>
-                <description xml:lang="x-default">clientNumber</description>
-                <type>string</type>
-                <localizable-flag>false</localizable-flag>
-                <mandatory-flag>true</mandatory-flag>
-                <externally-managed-flag>false</externally-managed-flag>
-                <min-length>0</min-length>
-            </attribute-definition>
-            <attribute-definition attribute-id="phoneNumbers">
-                <display-name xml:lang="x-default">Phone Numbers</display-name>
-                <type>set-of-string</type>
-                <localizable-flag>false</localizable-flag>
-                <mandatory-flag>false</mandatory-flag>
-                <externally-managed-flag>false</externally-managed-flag>
-            </attribute-definition>
-            <attribute-definition attribute-id="productName">
-                <display-name xml:lang="x-default">productName</display-name>
-                <type>string</type>
-                <localizable-flag>false</localizable-flag>
-                <mandatory-flag>false</mandatory-flag>
-                <externally-managed-flag>false</externally-managed-flag>
-                <min-length>0</min-length>
-            </attribute-definition>
-        </attribute-definitions>
-        <group-definitions>
-            <attribute-group group-id="Product-Subscribe-Notification">
-                <display-name xml:lang="x-default">Product-Subscribe-Notification</display-name>
-                <attribute attribute-id="productID"/>
-                <attribute attribute-id="productName"/>
-                <attribute attribute-id="phoneNumbers"/>
-            </attribute-group>
-        </group-definitions>
-    </custom-type>
-
 </metadata>
